extract components to index and fix nav tab protected areas link
