--- app/layout/scenarios/edit/planning-unit/adjust-planning-units/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
