--- app/layout/scenarios/edit/planning-unit/adjust-planning-units/tabs/index.ts
@@ -1,2 +1,0 @@
-export type { PlanningUnitTabsProps } from './component';
-export { default } from './component';
