--- app/layout/scenarios/edit/planning-unit/cost-surface/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
