--- app/layout/scenarios/edit/planning-unit/protected-areas/pa-selected/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
