--- app/layout/scenarios/edit/solutions/overview/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
