--- app/layout/scenarios/edit/solutions/overview/list/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
